Dont wipe cloud01 hosts
This will prevent wiping hosts if they are reclaimed and
we still need the data from them.

Fixes: https://github.com/redhat-performance/quads/issues/165

+++ bin/move-and-rebuild-host.sh
@@ -36,6 +36,7 @@ untouchable_hosts=${quads["untouchable_hosts"]}
 ipmi_username=${quads["ipmi_username"]}
 ipmi_password=${quads["ipmi_password"]}
 ipmi_cloud_username_id=${quads["ipmi_cloud_username_id"]}
+spare_pool_name=${quads["spare_pool_name"]}
 
 [ ! -d $lockdir ] && mkdir -p $lockdir
 
@@ -149,31 +150,32 @@ ipmitool -I lanplus -H mgmt-$host_to_move -U $ipmi_username -P $ipmi_password us
 #### BEGIN FOREMAN REBUILD
 
 if $rebuild ; then
+  if [ $new_cloud != "cloud01" ]; then
 
-  # first ensure PXE enabled on the host .... for foreman
-  $bindir/pxe-foreman-config.sh $host_to_move
+    # first ensure PXE enabled on the host .... for foreman
+    $bindir/pxe-foreman-config.sh $host_to_move
 
-  # also determine whether or not to leverage post snipper for PXE disablement
-  $bindir/pxe-director-config.sh $host_to_move $new_cloud
+    # also determine whether or not to leverage post snipper for PXE disablement
+    $bindir/pxe-director-config.sh $host_to_move $new_cloud
 
-  # either puppet facts or Foreman sometimes collect additional interface info
-  # this is needed sometimes as a workaround: clean all non-primary interfaces previously collected
-  skip_id=$(hammer host info --name $host_to_move | egrep -B 3 "nterface .primary, provision" | grep Id: | awk '{ print $NF }')
-  TMPIFFILE=$(mktemp /tmp/IFXXXXXXX)
-  hammer host info --name $host_to_move > $TMPIFFILE
+    # either puppet facts or Foreman sometimes collect additional interface info
+    # this is needed sometimes as a workaround: clean all non-primary interfaces previously collected
+    skip_id=$(hammer host info --name $host_to_move | egrep -B 3 "nterface .primary, provision" | grep Id: | awk '{ print $NF }')
+    TMPIFFILE=$(mktemp /tmp/IFXXXXXXX)
+    hammer host info --name $host_to_move > $TMPIFFILE
 
-  # remove extraneous interfaces collected prior to previous host usage
-  for interface in $(grep Id $TMPIFFILE  | grep ')' | grep -v $skip_id | awk '{ print $NF }') ; do
-      hammer host interface delete --host $host_to_move --id $interface
-  done
-  rm -f $TMPIFFILE
+    # remove extraneous interfaces collected prior to previous host usage
+    for interface in $(grep Id $TMPIFFILE  | grep ')' | grep -v $skip_id | awk '{ print $NF }') ; do
+        hammer host interface delete --host $host_to_move --id $interface
+    done
+    rm -f $TMPIFFILE
 
-  # perform host rebuild, in future the OS here should be a variable, fix me.
-  hammer host update --name $host_to_move --build 1 --operatingsystem "RHEL 7"
-  ipmitool -I lanplus -H mgmt-$host_to_move -U $ipmi_username -P $ipmi_password chassis power off
-  sleep 30
-  ipmitool -I lanplus -H mgmt-$host_to_move -U $ipmi_username -P $ipmi_password chassis power on
-
+    # perform host rebuild, in future the OS here should be a variable, fix me.
+    hammer host update --name $host_to_move --build 1 --operatingsystem "RHEL 7"
+    ipmitool -I lanplus -H mgmt-$host_to_move -U $ipmi_username -P $ipmi_password chassis power off
+    sleep 30
+    ipmitool -I lanplus -H mgmt-$host_to_move -U $ipmi_username -P $ipmi_password chassis power on
+  fi
 fi
 
 #### END FOREMAN REBUILD
